Refactor the usock code so it can be re-used for a new pmix server support library as well as the client. Introduct a server support library based on a function-shipping model that hides the PMIx internal communications from the server as well as the client. Thus, servers will now only deal with function calls that mirror the client APIs, with the server library acting to "ship" the function call across the two processes.
Update the test programs for the new library - added pmix_test.c as the revised server-side test code. Only a few functions have been tested, and I still need to implement support for a number of the APIs.

However, this commit should help to introduce the changes so the test codes can be properly updated.