Changes to collective logic:
1. Tracker is uniquely identified by the set of process AND collective
type.
2. Each node controls consistency of the data_collection flag for local
PMIx_Fence calls. In case of inconsistency the data collection flag will
be set to invalid to let other servers know that some server found error
in Fence signature. Final value of the data collection flag is appended
to the node-wide bucket.
3. At Fence finalization, data collection flags from all servers are
received and analyzed against local flag setting. This is done to make
sure that all servers will be aware that error was detected.