Refactor the code base into a single library. Unfortunately, we do have cases (e.g., OMPI) where processes need to be both server -and- client (e.g., the ORTE daemons). This leads to duplicate symbols unless we combine everything into a single library.
The overlapping functionality was in PMIx_Get and PMIx_Store_internal, and those are now resolved. However, we still need to resolve what to do when a dual-purpose process wants to publish or perform other client-level operations. While some of those may not be permitted, some are - e.g., publishing to a global server. This remains TBD.