Cleanup the direct modex operations
Resolve a race condition as we cannot know if a proc is local or remote until after all clients have been registered by the host RM. Add a lot of comments to the code flow, and remove compound if statements, to clarify the decision path. Get all the different code paths to agree on the format of the packed data returned by various modex operations.