Prototype for new tools connection mechanism
Add notes (marked by IBM) to the files explaining where code needs to be implemented, and giving guidance on what needs to be done.

Add the PMIx_Query_info(_nb) APIs to the client library - need to be implemented!

Add the query callback to the server module - needs to be implemented!

Resolve conflicts