client: skip connect/auth if PMIX_SERVER_FD is set
PMIx_Init() uses the file descriptor set in PMIX_SERVER_FD
if set in the environment.  Otherwise, connect to the UNIX
damain socket extracted from PMIX_SERVER_URI as before.

PMIX_SERVER_FD is then unset to prevent it being used again,
for example if inherited by a child process.

The init-auth handshake is also skipped in this case.