client: drop access(2) check in PMIx_Init()
Drop the access(2) check on the UNIX domain socket in PMIx_Init().
This check should only be done if connect(2) will be called.
Since connect(2) will also fail with an appropriate error if the
socket does not exist, it seemed simplest to just remove the check.

One minor change: we previously returned PMIX_ERR_NOT_FOUND if the
access failed, but return PMIX_ERR_UNREACH if connect fails.
Also, connects are retried, a fixed number of times, with no delay
bewteen attempts.