Simplify the PMIx_get processing:
If the key starts with “pmix”, then it is a PMIx-namespace key and can only be in the data provided at startup (which is stored in the “internal" hash table). Otherwise, it is data that was PMix_Put and is in the “modex” hash table. PMIx job-level data is stored in the “internal” hash table against rank=PMIX_RANK_WILDCARD.

This reduces all calls to PMIx_Get to a single hash-table lookup

Fix the problem of a proc asking for job-level data from another nspace by ensuring that the server detects the situation, and then returns the nspace blob for the other nspace.

Backport of Coverity warning fixes

Address review line comments from @artpol84.

Backport changes from OMPI

free(nspace) in job_data that is read and allocated but not used.  Still needs to be freed
even though we don't use it.

Fix two cases of pack being called with the wrong pointer type

When reading a namespace blob from a namespace other than our own, we still have to read
the nspace name off first and then we need to actually check if the blob contains what
we are looking for.

Update comments on keys per telecon

Check incoming keys to PMIx_Put for namespace reservation reject if they violate the PMIx reservation

Remove generated file

Add missing file back

Reset common file