Fix a bus error on reading dstore data segments
sparc64 (and some other architectures?) needs 64-bit alignment when
accessing 64-bit value but each kval entry in dstore data segments
is not necessarily aligned to 64-bit (or even 32-bit?). So we must
use `memcpy` instead of assignments when reading `size_t` values
(size of value and offset in extension slots).