dstor: fix PMIx_Get to request job-data for "pmix.*" keys
PMIx_get for all the requested keys it uses `fetch` with `WILDCARD`,
this leads to loss time dstor if the key is not a job-info.
This commit checks the key name is similar to "pmix.*" before to get
job-data keys by dstor, in this way it reduces the PMIx_Get time
in dstor case.