Add the ability to compress/decompress long strings per suggestion by @hjelmn. We can't do it exactly as he suggested as that approach only dealt with pack/unpack of buffers for communication. Our problem is more around the storage size, not the size of our messages. So check the length of strings that are about to be stored, and compress them if larger than a limit. Mark such strings with their decompressed size and a new PMIX_COMPRESSED_STRING data type so we know how to decompress them before delivery.
Also, remove the HWLOC support from PMIx. We really don't want to store the actual topology tree as (a) it is quite large on some architecures, and (b) the pointers in that tree require that we store a full copy of the tree in each client process. Instead, we will store the HWLOC topology xml string, properly compressed, and provide that back to the caller when requested. The caller will then have to "load" the topology tree from the xml string, free the string, and "destroy" the tree when they are done with it.

Fixes #204

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

Minor cleanup

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

More minor cleanups

Signed-off-by: Ralph Castain <rhc@open-mpi.org>