dstore: Fixed dstor bug case of key overwrite.
Data corruption was observed when a key was stored second time (forcing
data overwrite). In case of overwrite, the old record is invalidated
with "INVALIDATED" string written into its key name field. The problem
was observed when space insufficient for storing "INVALIDATED" string
was reserved for the key name field at the original key submission.
The fix provides minor refactoring of the dtore record formats to
flexibly address this issue.

Signed-off-by: Boris Karasev <karasev.b@gmail.com>