dstore: Fix significant performance degradation.
Accidentally loop that was submitting jobinfo data was duplicated
in two nested functions. As the result computational complexity
rised from O(n) to O(n^2) which caused significant performance
hit at 8k node level (thanks to @hjelmn for noticing and doing
initial analysis!)
This bug wasn't visible at dstore data level because if for the key
being replaced the size of the data was the same - key was simply
rewritten.

Signed-off-by: Artem Polyakov <artpol84@gmail.com>