dstore: Fix rare memory corruption
Thanks to @hjelmn for this discovery and invaluable help with debugging.

How to reproduce:
The error was a coincidence of 2 events:
* Previous rank data was completed and we are switching to a new rank.
* Previous rank together with extension slot occupied amount of space
exactly equal to  the segment size.

What was the root cause:
`(global_offset % data_segment_siza) == 0` is an acceptable case that
wasn't handled properly. Offset was turning into 0 confusing segment
boundary check logic.

The fix:
As global_offset as returned by `get_free_offset` can not be 0 we need
to check if `offset = global_offset % _data_segment_size` is turning
to a zero and consider that this is a sign that we need to create a new
segment.

Signed-off-by: Artem Polyakov <artpol84@gmail.com>