Check for failure to find any usable components in frameworks that require at least one, print an informative error message, and return an error.
If the user provides us with a PMIX_INSTALL_PREFIX that points to an unusable location (either nonexistent or not available to us), or none of the default search locations exist (because the installation was moved), then provide a nice message so they know what is wrong and return an error.

Allow the path search to continue beyond the first failed element.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>