jobinfo: avoid duplication of job information
This fix avoids double unpacking of the jobinfo array, which
occurs in nspace register, and then when clients are connected.
To provide job data to clients the local cache is used instead
of unpacking data from jobinfo array. Since the jobinfo array
is not needed now so it has been removed.