Fix the direct modex operations when gds=hash
When requesting modex data without it having already been collected, the client issues a request to the server for data "put" by the given rank. Since the server always returns all data for that proc, we don't pass the key being requested. Thus, the server operates with a NULL key for the request.

When a proc "commits" its modex data, the server stores it in different locations based on its scope. Local scope data goes into the GDS component for that peer, but remote scope data is stored under the server's GDS component. This is done to allow proper response to remote vs local procs - i.e., local procs don't see remote scope data, and vice versa.

The GDS/hash component is working correctly. When a request for data arrives, we look in the appropriate place based on whether the requested proc is local or remote. If we find it, we return that data. If we don't find it, then we create a tracker and wait for it to appear.

The GDS/ds12 component, however, is NOT working correctly. Because a NULL key is provided, it only looks to see if the job-level data is available, and then immediately returns with SUCCESS and the shared memory location. This is okay IF the requested proc has already "committed" its data, but causes the client proc to generate an error if the requested proc has NOT provided its data yet.

This is seen when running the following test in the test/simple directory:

$ ./simptest -n 2 -e ./simpdmodex

Signed-off-by: Ralph Castain <rhc@open-mpi.org>