General cleanup of tool and failure code paths
Protect receive callbacks from NULL and/or empty buffers as this
can occur when the peer on a connection disappears.

Fix tool connection search so it properly descends into the directory
tree while searching for the server's contact file.

Fix store_local so it doesn't reject a new nspace as that can happen
when working with tools

Ensure we _always_ complete PMIx_Finalize - don't return if something
goes wrong in the middle of the procedure

Signed-off-by: Ralph Castain <rhc@open-mpi.org>