Resolve dmodex of job-level info
When a process requests job-level info for a namespace unknown to the PMIx server, the server generates a dmodex request to its local host. The request specifies the target nspace, but with a WILDCARD rank since the data being requested is from the job-level info. At least one way the host can respond is to register the nspace using the PMIx server's register nspace API. However, this results in the data being "cached" in the server's GDS module, not in that of the requesting client. The solution to that problem lies in retrieving the data from the server's cache and then transferring it to the requesting nspace's GDS module.

An alternative response is for the host to simply provide the data in the dmodex response callback. In this case, we need to ensure that it gets stored in the PMIX_INTERNAL scope as that is where we look for job-level info. Support that option as well since there is no requirement to use the other method.

Also included a few minor cleanups that cropped up when compiling and trying to commit (e.g., whitespace)

Signed-off-by: Ralph Castain <rhc@open-mpi.org>

(cherry picked from commit 1207023c664ffe5b6c3cd101fceef425944979b8)