Prototype implementation of RFC0022
Environmental Parameter Directives for Applications and Launchers
It is sometimes desirable or required that standard environmental variables (e.g., PATH, LD_LIBRARY_PATH, LD_PRELOAD) be modified prior to executing an application binary or a starter such as mpiexec - this is particularly true when tools/debuggers are used to start the application. This RFC defines new attributes to support these needs.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>