Properly ensure that resources are returned at child end
Resources are allocated prior to knowing the type of application being executed. If the app never calls PMIx_Init, then no connection is made to the server and no pmix_peer_t object is created for it. This also means that PMIx_Finalize will never be called.

Together, this caused the client_finalized and app_finalized routines to fail to be called, thus leaking resources. Move those calls to the deregister_client function as the host RM is _always_ required to call that upon process termination, regardless of the type of application.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>