dstore: Avoid extra extension slots
Extra extension slots were placed when the key-values were
passed individually. Examples are:
- direct modex
- local keys at the commit time.

We need to place extension slots in order to make sure that
they are separating keys from different ranks (and thus can
provide the "next" link if new keys are to be appended).

This solution detects unneeded extension slot when keys from
the same rank are being placed. These unneeded extension slots
are removed prior to placing a new key.

Signed-off-by: Artem Y. Polyakov <artpol84@gmail.com>