Convert the code base to use pmix_strncpy
We have been seeing lots and lots of warnings appearing on certain
combinations of compiler and environment due to unused return value from
strncpy and the possibility that the resulting string could fill the
provided array without leaving room for a NULL terminator. In addition,
the security risk from non-NULL-terminated strings being input by the
user has been flagged.

Therefore, transition all uses of "strncpy" to a new "pmix_strncpy"
function that resolves the warnings and closes the security gap.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>