Add a timer to avoid memory leaks
Put the tracker in a LONG timeout in case the host never does call
back with the tracker after a local participant fails. This is a
minor detail, but for long-lived daemons we would prefer to avoid
memory leaks (albeit that this one is small).

Signed-off-by: Ralph Castain <rhc@open-mpi.org>
(cherry picked from commit 6d6193b8b4d8843673c94f5db56bcf290b0ed869)