Fix race conditions in tracked requests
When a process that is participating in a tracked request (e.g., collectives or direct modex) fails, a race condition exists between the lost-connection procedure and the host callback. Resolve this by checking and removing the proc from the trackers, and check the lists upon host callback prior to removing them from the list.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>