Attempt to address the case where data isn't found
If the server is asked for a piece of information that isn't immediately
available, we create a request tracker and add the requesting peer to
it. The precise return status, however, must be carefully controlled to
ensure that the peer object gets released when it should, and doesn't
get release twice.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>