Overhaul the event caching system
Events were being cached forever, leading to a flood hitting a newly
connected client that sometimes overwhelmed the system. There is likely
still something not quite right on the client side of the event
notification system as it shouldn't be getting overwhelmed. However,
cleaning up the system a bit won't hurt and we can then take a deeper
dive into the client side code.

Ensure that all cached objects (IOF requests, event registrations,
notifications) get purged when a proc finalizes or abruptly terminates,
and when nspaces are deregistered.

Signed-off-by: Ralph Castain <rhc@open-mpi.org>