Pass the tool's cmd line to the server
Ensure the server knows what the tool's cmd line was so that it can provide it in response to queries. Add an PMIX_QUERY_NAMESPACE_INFO attribute so we can return more than just a simple list of namespaces - e.g., to include the cmd line executing in the namespace.

Signed-off-by: Ralph Castain <rhc@pmix.org>