Implement support for class-based info arrays
The updated PMIx v3.1 Standard defines new attributes for passing arrays
of data for the various levels (job, app, node). This was done in order
to simplify the data being passed to register_nspace, and to support
passing of data that used the same attribute key for each level - e.g.,
the PMIX_NODE_SIZE for multiple nodes. This PR implements at least the
job and app arrays - the node array support remains TBD.

Signed-off-by: Ralph Castain <rhc@pmix.org>