Continue development of Python bindings
Fill in the server class APIs and begin development of the locks and callback functions

Resolve conversion of input dictionaries to pmix_info_t arrays by
executing it inside the object instead of trying to do it externally and
return the array. Extend the server.py example to fork/exec the
client.py test code and get the combination working.

Signed-off-by: Ralph Castain <rhc@pmix.org>