Pick up additional changes to the failed proc code path
Track if the collective request has been posted to the host. If it has,
then don't remove it from the tracker list and release it - wait for the
host to return. If not, then don't let it go to the host.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit a3dc09d0fb672ceace330a9e4b4dc2f6222b20e0)