Replicate the tests in current Mellanox Jenkins
Per chat with @jsquyres, use autogen.pl to assemble a file containing
the path names of all components in the code base, then let configure
AC_SUBST that string into the run_tests.pl for us. We then let Perl push
the result into the environment as an MCA parameter telling the system
where to find all the component libraries.

Signed-off-by: Ralph Castain <rhc@pmix.org>