mca base: warn if dir component_dir path fails to add
When pmix_mca_base_component_path is processed, it is not an error
(i.e., we should not abort in the middle) if one of the directories
fails to add.  Instead, we should warn the user that the directory
failed to add, but then keep processing all other directories listed
in the path.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit 1019712a9d99adb686a02c28c9b28549a063666c)