Fix the tests in the test subdir
Two errors were found. First, the return code when data isn't found can
be either PMIX_ERR_NOT_FOUND or PMIX_ERR_PROC_ENTRY_NOT_FOUND - there is
no guarantee which might be returned. Only guarantee is that it will not
be PMIX_SUCCESS. Thus, the test has to be a little more flexible when
checking for error codes.

Second, when multiple servers are trying to mimic a multi-node setup,
each server is required to provide a proc map of the entire job - not
just the clients local to it.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 96c8765f636a64ea2d9759af0b49622919513fa4)