Enable launcher tools to fork/exec local children
Tools such as debuggers need the ability to execute a local fork/exec of
a subprocess - e.g., a launcher such as "prun" - when operating in an
environment where there is no PMIx server available. When doing so, the
"child" retains the same nspace as the tool and we simply increment the
rank for each child. Note that this feature is ONLY available when the
tool is not connected to a server or else tracking of proc IDs would
become confused.

Implementing this exposed several errors involving assumptions about
tool-server connectivity that had to be fixed.

Signed-off-by: Ralph Castain <rhc@pmix.org>