Properly handle rendezvous files for launchers
Check the environment for PMIX_LAUNCHER_RENDEZVOUS_FILE, allowing a
provided info key to override any env value. If we are a launcher and
the file already exists, then assume this is where we find the info to
contact our server. If it doesn't exist, then assume this is where we
are to put our info.

Signed-off-by: Ralph Castain <rhc@pmix.org>