Provide proc data in cases where host does not
Enough information is in the node and proc maps to
compute most (if not all) of the data usually provided
in the PMIX_PROC_DATA entries. In the case where those
entries are not provided by the host, compute and store
the information based on the node and proc map data.

Signed-off-by: Ralph Castain <rhc@pmix.org>