Provide proc data in cases where host does not
Enough information is in the node and proc maps to
compute most (if not all) of the data usually provided
in the PMIX_PROC_DATA entries. In the case where those
entries are not provided by the host, compute and store
the information based on the node and proc map data.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 2156e7f7134374293261426a85c037e38359f6ec)