Fix xversion spawn test
Although the number of argvs is technically an int32, we have
to pack/unpack it as a generic int due to a typo in earlier
release series. This preserves the ordering of bytes in the
packed buffer as it includes a tag indicating the actual size
of the value. No harm is done as generic int is equivalent to
int32 on all current systems - just something to watch out
for in the future should someone someday change the size of "int"

Signed-off-by: Ralph Castain <rhc@pmix.org>