Complete implementation of network coord support
Provide a more flexible definition for pmix_coord_t that supports
arbitrary dimensionality and offers multiple points-of-view for the
fabric. Implement the pnet/test component to provide coordinates based
on an MCA parameter definition for planes, switches, and nodes. Update
the simple tests to exercise it and display the result.

Signed-off-by: Ralph Castain <rhc@pmix.org>