Move the testing of coord capabilities into separate test
Current xversion is attempting to use the simptest and simpclient from
the branch's test subdirectory. However, the versions in other branches
may be incompatible if we update these to support the latest features.

We need to fix this in our testing infrastructure, but for now, just
push the new feature tests into their own code.

Signed-off-by: Ralph Castain <rhc@pmix.org>