Be more restrictive on the minimum Python requirement
The Autoconf test searches the system for a Python of at least the
specified version. However, that does NOT mean that the active version
of Python met that requirement - it only means that the specified
version is somewhere on the system. We actually require that the active
version meet the minimums, so test that more directly.

Signed-off-by: Ralph Castain <rhc@pmix.org>