Resolve get of proc-specific job-level info from another nspace
The current code focused solely on retrieving modex info for procs where
the "get" request goes up to the PMIx server as there was an assumption
that job-level information would already be available. However, when
asking for information on a different nspace, the job-level info may be
available at the server, but not to the requesting process.

We had a code path to return the job-level info from the other nspace,
but that didn't include the proc-specific portion of the job-level info.
In other words, we returned things like PMIX_UNIV_SIZE but did not
return PMIX_HOSTNAME for each proc in the other nspace. This change
fixes that gap.

As a note for the future: it would be simpler if we just provided a
"hook" to the other proc's shmem region. However, that would require
some additional code to avoid leaving a proc with a stale handle to the
other nspace's shmem region once that nspace goes away.
Signed-off-by: Ralph Castain <rhc@pmix.org>

(cherry picked from commit 6e3ed184cb0a2a2bc1d4563bdb0b93a3e8d7996f)