Try to fix the "test" area
The tests do not correctly compute the node and proc maps when running
with multiple servers - they only compute those values for the local
server. This wasn't detected in the past because we weren't using that
info. However, now that we are correctly tracking info by node, this
causes the tests to become confused and fail.

I was unable to get the tests to correctly compute the pattern of proc
assignments for multiple servers, so just tell the tests to only run one
server until someone with more knowledge of this code can fix the
problems.

Signed-off-by: Ralph Castain <rhc@pmix.org>