Continue to work around the dstore's limitations
There are additional places where we need to get node or app-specific
info, so test for those keys and redirect to the hash component when
found. Note that this is impacting overall latency as these checks must
be done in the critical path prior to looking at the dstore.

Allow the server to pass its hostname to be used by its clients so that
a test can establish a "fake" hostname.

Signed-off-by: Ralph Castain <rhc@pmix.org>
(cherry picked from commit 8036edf2d313ea7d6d9676a1a3fdc91defc62db4)
Signed-off-by: Ralph Castain <rhc@pmix.org>