Deal with servers using older versions
Servers prior to v3.1.5 don't properly track information such as
local_peers on a per-node basis. So ensure we try to access peer info in
a manner consistent with the server's version.

Also, the test servers in older versions just return "not found" for
direct modex requests. This means that resolve_peers is guaranteed to
fail for requests of information about other namespaces, so don't
include such tests in that situation.

Signed-off-by: Ralph Castain <rhc@pmix.org>