Final changes to enable cycling of init/finalize
The PMIx library needs to be able to support multiple calls to
PMIx_Init/PMIx_Finalize as an app might cycle thru these calls prior to
entering a PMIx-enabled library. We could set a rule prohibiting calling
init if the library has been completely finalized (i.e., if the number
of calls to finalize equals the number of calls to init, thus driving
the refcount to zero and causing the library to fully close down), but
that seems overly restrictive.

Signed-off-by: Ralph Castain <rhc@pmix.org>