Complete debugging of the collective tracker
There remain some issues about detecting local vs non-local procs when
setting up the tracker. I've resolved the ones surfaced by multi-node
testing using PRRTE with both simple PMIx clients and the "double-get"
test case.

For the latter, I cannot find any solution that is totally transparent
to the user. The local daemon has no way of knowing if the remote daemon
has received more info or updated info. So we cannot know if we should
go ask for the info or not - we have a blob and that is all we can
return.

Only solution I can find is to define a new attribute
(PMIX_GET_REFRESH_CACHE) that tells the local server to go get
potentially updated remote info. Note that this in itself doesn't mean
that we will wait for that info to appear - that remains for another
day.

Signed-off-by: Ralph Castain <rhc@pmix.org>