Complete debugging of the collective tracker
There remain some issues about detecting local vs non-local procs when
setting up the tracker. I've resolved the ones surfaced by multi-node
testing using PRRTE with both simple PMIx clients and the "double-get"
test case. This is a subset of those changes as the broader ones failed
to pass CI - hopefully, this minimum subset does and then I'll try again
with the remainder.

For this subset, we use the fact that we are being told both the number
of local procs and the total number of procs in a namespace upon call to
"register_nspace". If we are in a collective and have not yet received
the "register_nspace" call, then we defer processing of the collective
until we do.

Signed-off-by: Ralph Castain <rhc@pmix.org>