Properly support dmodex requests
A dmodex operation is triggered whenever the key being requested is not
found in the local cache. Include the specified key in the request
passed from the client to the local server. Have the server check for
that key in its own storage - if it isn't present, then pass the request
to the host thru the dmodex interface with the key being identified by
including the PMIX_REQUIRED_KEY attribute in the info array.

Also cleanup some of the PMIx client get that was obscured by the
compensating errors in the tests. Part of the problem here is that the
tests only verify that the end result was okay - they don't check to see
if the code obtained the result correctly. For example, we might well
wind up doing a dmodex instead of finding the info locally in the dstore
(which is what was happening here). Need to ponder a better way of
testing the code.

Fix the test/job-fence code so it doesn't wait forever on PMIx_Get of a
non-existent key.

Signed-off-by: Ralph Castain <rhc@pmix.org>