Properly support dmodex requests (#1611)
* Properly support dmodex requests

A dmodex operation is triggered whenever the key being requested is not
found in the local cache. Include the specified key in the request
passed from the client to the local server. Have the server check for
that key in its own storage - if it isn't present, then pass the request
to the host thru the dmodex interface with the key being identified by
including the PMIX_REQUIRED_KEY attribute in the info array.

Also cleanup some of the PMIx client get that was obscured by the
compensating errors in the tests. Part of the problem here is that the
tests only verify that the end result was okay - they don't check to see
if the code obtained the result correctly. For example, we might well
wind up doing a dmodex instead of finding the info locally in the dstore
(which is what was happening here). Need to ponder a better way of
testing the code.

Fix the test/job-fence code so it doesn't wait forever on PMIx_Get of a
non-existent key.

Signed-off-by: Ralph Castain <rhc@pmix.org>

* Deal with the differences between versions

Early versions don't know to provide the key when sending "get" to the PMIx server library

Signed-off-by: Ralph Castain <rhc@pmix.org>

* Handle v2.2 oddities

v2.2 still requires that the dstore contain keyvals where the key is the
hostname and the value is an info array for that host. Check for that
version and accommodate it.

Silence a couple of nit warnings in simptest

Signed-off-by: Ralph Castain <rhc@pmix.org>