Allow unspecified node/app queries to fall thru
We populate the dstore with the node/app values pertinent to our local
node/app. If the call to "get" does not specify a different node/app,
then assume they want our local values and allow the dstore to service
the request. This avoids a query being sent to the local server for the
info that can subsequently be retrieved from the hash component.

Signed-off-by: Ralph Castain <rhc@pmix.org>