Silence errors coming out of libtool config check
When running on a Mac, Apple's libtool does not support the "--config"
option. This leads to configure generating an error that can be
misleading. We _could_ use "glibtool" to perform the check, but we are
actually using Apple's libtool to do the build, so best to just silently
return the error

Signed-off-by: Ralph Castain <rhc@pmix.org>